Refactor: 로그인 회원가입페이지의 이벤트 발생시 DOM 조작요소 모듈화
- authDOMHandler.js : 로그인 회원가입 페이지의 DOM 조작함수 모듈
- authModule에서 일관적인 이름을 위해 authEventHandler로 변경
