+++ js/signin.js
@@ -1,5 +1,6 @@
 import { submitBtn, emailInput, passwordInput } from "./modules/domSelectors.js";
-import { authEvent, specifyWarningPosition } from "./modules/authModule.js";
+import { authEvent } from "./modules/authEventHandler.js";
+import { specifyWarningPosition } from "./modules/authEventHandler.js";
 import { verifyLoginCredentials } from "./modules/verifyUser.js";
 
 authEvent();
